+++ src/main/java/com/devpeople/bapsim/global/exception/AddressNotFoundException.java
@@ -1,7 +1,7 @@
 package com.devpeople.bapsim.global.exception;
 
 public class AddressNotFoundException extends RuntimeException {
-    public AddressNotFoundException(Integer id) {
+    public AddressNotFoundException(Long id) {
         super("주소를 찾을 수 없습니다. id=" + id);
     }
 }
